finish chapter 4

+++ part-4/blog-list/controllers/blogs.js
@@ -1,24 +1,50 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
 })
   
 blogsRouter.post('/', async (request, response) => {
-    const title = request.body.title
-    const url = request.body.url
+    const { title, url } = request.body
 
     if (!title || !url) return response.status(400).end()
 
-    const newBlog = new Blog(request.body)
+    const decodedToken = request.token
+        ? jwt.verify(request.token, process.env.SECRET)
+        : undefined
+    if (!(decodedToken && decodedToken.id)) return response.status(401).json({ error: 'token invalid' })
+
+    const user = request.user
+
+    const newBlog = new Blog({
+        title: request.body.title,
+        author: request.body.author,
+        url: request.body.url,
+        likes: request.body.likes,
+        user: user.id
+    })
     const savedBlog = await newBlog.save()
+
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) return response.status(401).json({ error: 'token invalid' })
+
     const blog = await Blog.findByIdAndRemove(request.params.id)
+
+    const user = request.user
+    user.blogs = user.blogs.filter(blog => blog.id !== request.params.id)
+    await user.save()
+
     response.status(blog ? 204 : 404).end()
 })
